+++ extensions/oci/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Oracle Cloud Changelog
 
-## [] - {PR_MERGE_DATE}
+## [Perform Instance Actions] - {PR_MERGE_DATE}
 
+- Add instance actions:
+    - Reset
+    - Start
+    - Stop
 - Show a proper `Detail` view with information when "provider" fails. Previously, extension would crash (ref: Issue #).
+- Show instance "Created" `Accessory`
 
 ## [Initial Version] - 2024-12-20